add cluster word_cloud
